# Inline stage 1