--- a/src/jmh/java/org/simdjson/ParseAndSelectBenchmark.java
+++ b/src/jmh/java/org/simdjson/ParseAndSelectBenchmark.java
@@ -39,7 +39,7 @@ public class ParseAndSelectBenchmark {
             buffer = is.readAllBytes();
             bufferPadded = padded(buffer);
         }
-        System.out.println("VectorSpecies = " + StructuralIndexer.BYTE_SPECIES);
+        System.out.println("VectorSpecies = " + VectorUtils.BYTE_SPECIES);
     }
 
     @Benchmark