--- a/src/jmh/java/org/simdjson/Utf8ValidatorBenchmark.java
+++ b/src/jmh/java/org/simdjson/Utf8ValidatorBenchmark.java
@@ -1,7 +1,15 @@
 package org.simdjson;
 
 import com.google.common.base.Utf8;
-import org.openjdk.jmh.annotations.*;
+import org.openjdk.jmh.annotations.Benchmark;
+import org.openjdk.jmh.annotations.BenchmarkMode;
+import org.openjdk.jmh.annotations.Level;
+import org.openjdk.jmh.annotations.Mode;
+import org.openjdk.jmh.annotations.OutputTimeUnit;
+import org.openjdk.jmh.annotations.Param;
+import org.openjdk.jmh.annotations.Scope;
+import org.openjdk.jmh.annotations.Setup;
+import org.openjdk.jmh.annotations.State;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -11,6 +19,7 @@ import java.util.concurrent.TimeUnit;
 @BenchmarkMode(Mode.Throughput)
 @OutputTimeUnit(TimeUnit.SECONDS)
 public class Utf8ValidatorBenchmark {
+
     @Param({"/twitter.json", "/gsoc-2018.json", "/github_events.json"})
     String fileName;
     byte[] bytes;
@@ -24,7 +33,7 @@ public class Utf8ValidatorBenchmark {
 
     @Benchmark
     public void utf8Validator() {
-        Utf8Validator.validate(bytes);
+        Utf8Validator.validate(bytes, bytes.length);
     }
 
     @Benchmark