--- a/src/main/java/org/simdjson/JsonCharacterBlock.java
+++ b/src/main/java/org/simdjson/JsonCharacterBlock.java
@@ -1,8 +1,0 @@
-package org.simdjson;
-
-record JsonCharacterBlock(long whitespace, long op) {
-
-    long scalar() {
-        return ~(op | whitespace);
-    }
-}