--- a/src/main/java/org/simdjson/SimdJsonParser.java
+++ b/src/main/java/org/simdjson/SimdJsonParser.java
@@ -2,12 +2,10 @@ package org.simdjson;
 
 public class SimdJsonParser {
 
-    private static final int STEP_SIZE = 64;
     private static final int PADDING = 64;
     private static final int DEFAULT_CAPACITY = 34 * 1024 * 1024; // we should be able to handle jsons <= 34MiB
     private static final int DEFAULT_MAX_DEPTH = 1024;
 
-    private final BlockReader reader;
     private final StructuralIndexer indexer;
     private final BitIndexes bitIndexes;
     private final JsonIterator jsonIterator;
@@ -24,23 +22,20 @@ public class SimdJsonParser {
         jsonIterator = new JsonIterator(bitIndexes, stringBuffer, capacity, maxDepth, PADDING);
         schemaBasedJsonIterator = new SchemaBasedJsonIterator(bitIndexes, stringBuffer, PADDING);
         paddedBuffer = new byte[capacity];
-        reader = new BlockReader(STEP_SIZE);
         indexer = new StructuralIndexer(bitIndexes);
     }
 
     public <T> T parse(byte[] buffer, int len, Class<T> expectedType) {
-        stage0(buffer);
         byte[] padded = padIfNeeded(buffer, len);
-        reset(padded, len);
-        stage1(padded);
+        reset();
+        stage1(padded, len);
         return schemaBasedJsonIterator.walkDocument(padded, len, expectedType);
     }
 
     public JsonValue parse(byte[] buffer, int len) {
-        stage0(buffer);
         byte[] padded = padIfNeeded(buffer, len);
-        reset(padded, len);
-        stage1(padded);
+        reset();
+        stage1(padded, len);
         return jsonIterator.walkDocument(padded, len);
     }
 
@@ -52,25 +47,13 @@ public class SimdJsonParser {
         return buffer;
     }
 
-    private void reset(byte[] buffer, int len) {
-        indexer.reset();
-        reader.reset(buffer, len);
+    private void reset() {
         bitIndexes.reset();
         jsonIterator.reset();
     }
 
-    private void stage0(byte[] buffer) {
-        Utf8Validator.validate(buffer);
-    }
-
-    private void stage1(byte[] buffer) {
-        while (reader.hasFullBlock()) {
-            int blockIndex = reader.getBlockIndex();
-            indexer.step(buffer, blockIndex, blockIndex);
-            reader.advance();
-        }
-        indexer.step(reader.remainder(), 0, reader.getBlockIndex());
-        reader.advance();
-        indexer.finish(reader.getBlockIndex());
+    private void stage1(byte[] buffer, int length) {
+        Utf8Validator.validate(buffer, length);
+        indexer.index(buffer, length);
     }
 }