--- a/src/main/java/org/simdjson/StringParser.java
+++ b/src/main/java/org/simdjson/StringParser.java
@@ -9,7 +9,7 @@ class StringParser {
 
     private static final byte BACKSLASH = '\\';
     private static final byte QUOTE = '"';
-    private static final int BYTES_PROCESSED = StructuralIndexer.BYTE_SPECIES.vectorByteSize();
+    private static final int BYTES_PROCESSED = VectorUtils.BYTE_SPECIES.vectorByteSize();
     private static final int MIN_HIGH_SURROGATE = 0xD800;
     private static final int MAX_HIGH_SURROGATE = 0xDBFF;
     private static final int MIN_LOW_SURROGATE = 0xDC00;
@@ -30,7 +30,7 @@ class StringParser {
         int src = idx + 1;
         int dst = offset;
         while (true) {
-            ByteVector srcVec = ByteVector.fromArray(StructuralIndexer.BYTE_SPECIES, buffer, src);
+            ByteVector srcVec = ByteVector.fromArray(VectorUtils.BYTE_SPECIES, buffer, src);
             srcVec.intoArray(stringBuffer, dst);
             long backslashBits = srcVec.eq(BACKSLASH).toLong();
             long quoteBits = srcVec.eq(QUOTE).toLong();