--- a/src/test/java/org/simdjson/ObjectParsingTest.java
+++ b/src/test/java/org/simdjson/ObjectParsingTest.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.simdjson.TestUtils.toUtf8;
+import static org.simdjson.testutils.TestUtils.toUtf8;
 import static org.simdjson.testutils.SimdJsonAssertions.assertThat;
 
 public class ObjectParsingTest {