… src/main/java/org/simdjson/BlockReader.java
… src/main/java/org/simdjson/CharactersClassifier.java
… src/main/java/org/simdjson/JsonStringBlock.java
… src/main/java/org/simdjson/JsonStringScanner.java
… src/main/java/org/simdjson/Utf8Validator.java
… src/main/java/org/simdjson/VectorUtils.java
… src/test/java/org/simdjson/BenchmarkCorrectnessTest.java
… src/test/java/org/simdjson/BlockReaderTest.java
… src/test/java/org/simdjson/BooleanParsingTest.java
… src/test/java/org/simdjson/BooleanSchemaBasedParsingTest.java
… src/test/java/org/simdjson/CharactersClassifierTest.java
… src/test/java/org/simdjson/FloatingPointNumberSchemaBasedParsingTest.java
… src/test/java/org/simdjson/IntegralNumberSchemaBasedParsingTest.java
… src/test/java/org/simdjson/JsonStringScannerTest.java
… src/test/java/org/simdjson/NullParsingTest.java
… src/test/java/org/simdjson/NumberParsingTest.java
… src/test/java/org/simdjson/ObjectSchemaBasedParsingTest.java
… src/test/java/org/simdjson/StringParsingTest.java
… src/test/java/org/simdjson/StringSchemaBasedParsingTest.java
… src/test/java/org/simdjson/StructuralIndexerTest.java
… src/test/java/org/simdjson/TestUtils.java
… src/test/java/org/simdjson/Utf8ValidationTest.java
… src/test/java/org/simdjson/Utf8ValidatorTest.java
… src/test/java/org/simdjson/testutils/TestUtils.java
… src/test/java/org/simdjson/testutils/Utf8TestData.java