I'm manually inlining all vector operations. This is the result of a discussion on [Panama's mailing list](https://mail.openjdk.org/pipermail/panama-dev/2024-June/020445.html).

### 256-bit vectors

| Benchmark                                                                      | main (ops/s) | inline_stage1 (ops/s) |
|--------------------------------------------------------------------------------|--------------|-----------------------|
| ParseAndSelectBenchmark.countUniqueUsersWithDefaultProfile                     |      955.428 |               990.783 |
| ParseAndSelectBenchmark.countUniqueUsersWithDefaultProfile (padded)            |      983.696 |              1048.394 |
| SchemaBasedParseAndSelectBenchmark.countUniqueUsersWithDefaultProfile          |     1436.906 |              1566.293 |
| SchemaBasedParseAndSelectBenchmark.countUniqueUsersWithDefaultProfile (padded) |     1536.843 |              1707.806 |

### 512-bit vectors
| Benchmark                                                                      | main (ops/s) | inline_stage1 (ops/s) |
|--------------------------------------------------------------------------------|--------------|-----------------------|
| ParseAndSelectBenchmark.countUniqueUsersWithDefaultProfile                     |     1736.325 |              1838.197 |
| ParseAndSelectBenchmark.countUniqueUsersWithDefaultProfile (padded)            |     1851.136 |              1931.859 |
| SchemaBasedParseAndSelectBenchmark.countUniqueUsersWithDefaultProfile          |     2973.474 |              3146.485 |
| SchemaBasedParseAndSelectBenchmark.countUniqueUsersWithDefaultProfile (padded) |     3163.667 |              3512.966 |